requirements.txt
# Backend
fastapi
uvicorn
faiss-cpu
pypdf
transformers
sentence-transformers

# Frontend
streamlit
requests
python-dotenv
